integracion de rama con el proyecto actual

+++ Proyecto/Backend/ms-usuarios/src/main/java/cl/PetDate/ms_usuarios/repositories/UsuarioRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends MongoRepository<Usuario,Integer> {
+public interface UsuarioRepository extends MongoRepository<Usuario,Long> {
 
     Optional<Usuario> findByCorreo(String correo);
